spline #772 Addressing PRs comments. Split RESTClient contract and impl.
